Add working HyPro and CArl revisions to README.md
Previously I recommended the 17.10 releases because they enabled me to
build hpnmg itself and released versions seemed better to me than some
specific git commit.

However, some targets, for example testEvent, could not be built because
of linkage problems with libhpnmg.so. Somehow, this was related to the
libglpk dependency included in HyPro.

Since compiling against the HyPro master does not work because of moved
files and updated method signatures, we have to record a specific
revision in the installation instructions.