Fix region creation with normalized rv intervals
The dependency coefficients of the RVintervals and the ones in the
intervalBounds were sorted in local firing order of the location. This
resulted in wrong computations of the base region of a location.

This introduces a normalization like we already of for the
generalDependencies of the Event class: The coefficients of the normed
interval vectors are sorted in the usual global firing order.