Fix halfspace creation for continuous place levels
This adds the normalized version of the continuous marking to the PLT
like we already have for the RV intervals and event times.

Since using those did not fix intersectRegionForContinuousLevel(), I
created a new version of the method because I could not work out how it
was supposed to calculate the correct linear equations. To keep the
impact minimal and in case the previous implementation is actually
useful, I renamed it to legacyIntersectRegionForContinuousLevel().