Refactor Region class to support open facets
HyPro does not support open halfspaces and, thus, does not support
polytopes with open facets. However, these are needed to for model
checking the negation operator.

Previously, Region was a general purpose wrapper for a hypro polytope
but it made that polytope publicly available. Its new form, STDPolytope,
is supposed to model exactly what we expect of a polytope that is a
subset of an STD region: The first dimensions corresponding to random
variables, the final dimension corresponding to time and explicit
support for the operations required for the STD model checking:
- extracting a slice from the polytope at a specific point in time
- intersection
- set difference

The last operation results in open factes so the polytope is now an
encapsulated implementation detail in order to achieve the freedom to
simulate open facets.

This also creates a dedicated Simplex class for the integration
intervals and transformation function because these algorithms do not
make any assumptions about the meaning of the dimension. Also, that
we're now able to uphold the invariant, i.e. that the wrapped polytope
is a simplex, from the construction onward is a huge plus design-wise.