Improve compilation times for STDPolytope template
Put briefly, this is achieved by removing the template definitions from
the header and instead explicitly instantiating the template with the
needed types in an extra compilation unit.

Less briefly:
If we want users of the library to be able to choose the Numeric type
for STDPolytpe, the template definitions *need* to be in the header.
However, this will greatly increase compilation time during development
of hpnmg because for every change to the implementation, i.e. the
template definitions, every file including the header needs to be
recompiled.
Instead, the template file is commented out and the template is
explicitly instantiated in STDPolytope.cpp for the types we actually use
during development.