Use GMP rationals in RegionModelChecker
The are frequent cases where we seem to have an assertion failure within
HyPro that is caused by numerical imprecision: The HPolytope constructor
for a set of vertices asserts that the constructed HPolytope actually
contains the vertices.

This is a halfway solution towards converting the whole hpnmg project to
be generic over the used number type.