Fix ParametricLocation::getRVIntervalsNormed()
The normed version of getRVIntervals was supposed to offer two things:
1. Sort the coefficient vectors in a consistent way.
2. Return consistently many and consistently ordered intervals.

These were addressed by choosing a canonical global order: Find the
largest number of firings for every general transition independently
throughout the evolution of the model. Then, the `n_0` firings of the
0th transition make up the first `n_0` random variables. The `n_1`
firings of the 1st transition make up the next `n_1` random variables
and so on.

Point 2 was overlooked in the implementation and instead only the random
variables actually existing in the location were returned in the firing
order specific to that location.