Fix ProbabilityCalculator for degenerate polytopes
For polytopes of non-full dimension, the triangulation offered by CGAL
results in segmentation faults when reading the vertices. I did not find
a way around this because the vertex-iterator can not be inspected in
any way before dereferencing it.

Since the probability for such polytopes is zero* anyway, we can just
skip the whole calculation.

* Well, at least as long as we do not allow Dirac distributions.