Fix RV bounds for concurrent general transitions
Previously, the code simply added a dependency coefficient of 1 for the
current firing of a general transition to the left bound of every other
enabled general transition. However, this is wrong for example for three
concurrently enabled general transitions: The firing time of the third
transition depends only on that of the second because that already
includes the firing time of the first one.

Instead, simply setting the left bounds to the current value of the
general clock should do the trick: The clock states the time the
transition has been enabled for now. This, however, is exactly the
minimum value for the random variable.