Optimize evaluation of trivial formulae
A False-formula type is added to explicitly complement the True-type.
With these types, a formula can recursively be checked for triviality,
i.e. whether the formula evaluates to constantly true or false.

This allows to optimize conjunction, negation and until formulae.
Especially the "finally" operator, i.e. "true until", benefits from
this.