Optimize Until checking in RegionModelChecker
The check function `until()` returned a proper satisfaction set for a
region, i.e. all polytopes in the set were limited to the region. This
is correct and expected from normal users of the `until()` function.
However, when `until()` recursively invoked itself, the first thing it
did to the resulting polytopes was to extend them downwards. By adding
an extra wrapper for the recursion, this artifical extension-limit-extension
process is removed.