Fix halfspace creation for continuous place levels
For locations whose level is constant, i.e. the drift is zero and the
place's initial level does not depend on random variables, we cannot
create a normal halfspace because the normal vector would be the
zero-vector.
Instead, the actual level can simply be compared to the desired level
and return either the full region for the location or an empty region.