Refactor include directives
We had many files that relied on transitively included headers.
Many files, especially headers, included headers which they didn't use.
This has a considerable effect on compilation times because
recompilation of the whole project could be easily trigged by a change
to a seemingly innocent header that is transitively included everywhere.

This commit walks through all header files, purges almost all includes
and then adds only those that are necessary for symbols in the header.
Then, compilation is attempted and missing headers are added to the
implementation files instead of their header, if possible.

Additionally, the following order of includes is applied:
1. For implementation files: their respecitve header file.
2. All includes from hpnmg
3. Includes from third party libraries, grouped by library and roughly
   ordered descendingly by their abstraction level, e.g. more
   foundational libriaries are included later.
4. The most foundational, i.e. standard libraries, are included last.
While this order is not of vital importance, it increases the likelihood
of non-self-contained headers standing out early. For example, if
`<string>` is included after some library header that uses `std::string`
but does not include `<string>` itself, the compilation will fail early.