Added a Logging class
The class Logging can now be used to log messages on different
severity levels. Logs are written to std::out as well as
hpnmg.log file. For both targets different LogLevel can be used
which are specified at runtime using the file-loglevel or
stdout-loglevel program options. Logging is based on the
library carl.