Revised Logger integration
According to the comments in the merge request the Logger
no longer needs a Flag to be used. We decided the Logger
to be an essential part of the library. Including now is
implemented in the *.cpp files. Additionally the Logger
class now belongs to the path /util/logger to have a more
fine grained util overview.