Added classes for generating statistics
The Statistics.h Header-only class can now be used to generate
statistics about the execution of an hpnmg experiment. It implements
(1) a CounterRegistry to allow multiple different counters e.g.
to count the number of executions of a function and (2) a wrapper
definition to measure the execution time. Outputs are then saved
in a hpnmg-statistics.log file. The usage of the statistics can be
defined using the HPNMG_STATISTICS directive in the flag.h file.